Still WIP not compiling. Deleted the actors and moved all the futures code into RecordReceiver (to be splitted and renamed)
