+++ transformer/src/main/scala/uk/ac/wellcome/transformer/modules/StreamRecordProcessorFactoryModule.scala
@@ -2,7 +2,6 @@ package uk.ac.wellcome.platform.transformer.modules
 
 import java.util.{List => JList}
 import javax.inject.Singleton
-import scala.collection.JavaConverters._
 
 import com.amazonaws.services.dynamodbv2._
 import com.amazonaws.services.kinesis.clientlibrary.interfaces._
@@ -10,14 +9,13 @@ import com.amazonaws.services.kinesis.clientlibrary.lib.worker._
 import com.amazonaws.services.kinesis.model.Record
 import com.google.inject.Provides
 import com.twitter.inject.TwitterModule
-
-import akka.actor.ActorRef
 import uk.ac.wellcome.finatra.modules._
-import uk.ac.wellcome.platform.transformer.modules._
-import uk.ac.wellcome.models.ActorRegister
+import uk.ac.wellcome.transformer.receive.RecordReceiver
+
+import scala.collection.JavaConverters._
 
 class StreamsRecordProcessor(client: AmazonDynamoDB,
-                             receiver: Option[ActorRef]
+                             receiver: RecordReceiver
 ) extends IRecordProcessor {
 
   case class ExampleRecord(identifier: String)
@@ -37,20 +35,19 @@ class StreamsRecordProcessor(client: AmazonDynamoDB,
     checkpointer: IRecordProcessorCheckpointer
   ): Unit = {
     records.asScala.foreach { record =>
-      receiver.foreach(_ ! record)
+      receiver.receiveRecord(record)
     }
   }
 }
 
 class StreamsRecordProcessorFactory(
   dynamoClient: AmazonDynamoDB,
-  actorRegister: ActorRegister
+  recordReceiver: RecordReceiver
 ) extends IRecordProcessorFactory {
 
   override def createProcessor(): IRecordProcessor =
     new StreamsRecordProcessor(dynamoClient,
-                               actorRegister.actors
-                                 .get("kinesisDynamoRecordExtractorActor"))
+                               recordReceiver)
 
 }
 
@@ -61,7 +58,7 @@ object StreamsRecordProcessorFactoryModule extends TwitterModule {
   @Provides
   def provideStreamsRecordProcessorFactory(
     dynamoClient: AmazonDynamoDB,
-    actorRegister: ActorRegister
+    recordReceiver: RecordReceiver
   ): StreamsRecordProcessorFactory =
-    new StreamsRecordProcessorFactory(dynamoClient, actorRegister)
+    new StreamsRecordProcessorFactory(dynamoClient, recordReceiver)
 }
